+++ bitcoin/src/blockdata/script/borrowed.rs
